getAssertionIdError testcase completed. Added selection and try catch blocks to operation-id-service and web3-service

+++ src/service/operation-id-service.js
@@ -97,7 +97,9 @@ class OperationIdService {
     async removeOperationIdCache(operationId) {
         this.logger.debug(`Removing operation id: ${operationId} cached data`);
         const operationIdCachePath = this.fileService.getOperationIdDocumentPath(operationId);
-        await this.fileService.removeFile(operationIdCachePath);
+        if (await this.fileService.fileExists(operationIdCachePath)) {
+            await this.fileService.removeFile(operationIdCachePath);
+        }
         this.removeOperationIdMemoryCache(operationId);
     }
